add formatPartnersForExport
