+++ apps/web/app/(ee)/api/partners/export/route.ts
@@ -1,25 +1,13 @@
 import { convertToCSV } from "@/lib/analytics/utils/convert-to-csv";
+import { formatPartnersForExport } from "@/lib/api/partners/format-partners-for-export";
 import { getPartners } from "@/lib/api/partners/get-partners";
 import { getPartnersCount } from "@/lib/api/partners/get-partners-count";
 import { getDefaultProgramIdOrThrow } from "@/lib/api/programs/get-default-program-id-or-throw";
 import { withWorkspace } from "@/lib/auth";
 import { qstash } from "@/lib/cron";
-import {
-  exportPartnerColumns,
-  partnersExportQuerySchema,
-} from "@/lib/zod/schemas/partners";
+import { partnersExportQuerySchema } from "@/lib/zod/schemas/partners";
 import { APP_DOMAIN_WITH_NGROK } from "@dub/utils";
 import { NextResponse } from "next/server";
-import { z } from "zod";
-
-const columnIdToLabel = exportPartnerColumns.reduce((acc, column) => {
-  acc[column.id] = column.label;
-  return acc;
-}, {});
-
-const numericColumns = exportPartnerColumns
-  .filter((column) => column.numeric)
-  .map((column) => column.id);
 
 const MAX_PARTNERS_TO_EXPORT = 1000;
 
@@ -59,49 +47,7 @@ export const GET = withWorkspace(
       programId,
     });
 
-    const columnOrderMap = exportPartnerColumns.reduce((acc, column, index) => {
-      acc[column.id] = index + 1;
-      return acc;
-    }, {});
-
-    columns = columns.sort(
-      (a, b) => (columnOrderMap[a] || 999) - (columnOrderMap[b] || 999),
-    );
-
-    const schemaFields = {};
-    columns.forEach((column) => {
-      if (numericColumns.includes(column)) {
-        schemaFields[columnIdToLabel[column]] = z.coerce
-          .number()
-          .optional()
-          .default(0);
-      } else {
-        schemaFields[columnIdToLabel[column]] = z
-          .string()
-          .optional()
-          .default("");
-      }
-    });
-
-    const formattedPartners = partners.map((partner) => {
-      const result = {};
-
-      columns.forEach((column) => {
-        let value = partner[column] || "";
-
-        // Handle date fields - convert to ISO string format
-        if (
-          (column === "createdAt" || column === "payoutsEnabledAt") &&
-          value instanceof Date
-        ) {
-          value = value.toISOString();
-        }
-
-        result[columnIdToLabel[column]] = value;
-      });
-
-      return z.object(schemaFields).parse(result);
-    });
+    const formattedPartners = formatPartnersForExport(partners, columns);
 
     return new Response(convertToCSV(formattedPartners), {
       headers: {
